// src/components/atoms/AnimationLines/types.ts
export interface IWrapperType {
  zIndex?: number;
  intend: number;
  scale: number;
  sizes: {
    line1: {
      width: number;
      height: number;
    };
    line2: {
      width: number;
      height: number;
    };
    line3: {
      width: number;
      height: number;
    };
    line4: {
      width: number;
      height: number;
    };
  };
}